# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SPRITE-INDEX API",
    "version": "1.0.0",
    "description": "Contrat public des endpoints HTTP principaux."
  },
  "servers": [
    {
      "url": "https://sprite-index.com",
      "description": "Production"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session token"
      }
    },
...
  },
  "paths": {
...
    "/api/collection/{userId}": {
      "get": {
        "summary": "Lit une collection visible",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection"
          },
          "403": {
            "description": "Collection privée"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}